nsl_usb/sie_descriptor: Rework lookup, use self-descriptive structure in ROM
Descriptor lookup may take some time, but this avoids having a big
offset table implemented in LUTs.

USB2 mandates we are able to respond to control read within 50 ms. We
are still doing it in the back-to-back OUT (SETUP Data0 to IN token)
delay.

Also add support for one dynamic string descriptor. This is mostly
useful for serial numbers.