all: Rework clock+reset process modelization
One of the usual ways to model asynchronous reset is:

  if reset then
    ...
  elsif rising_edge(...) then
    ...
  end if;

In case we have some registers that are not reset in the reset clause (mostly
because they are not used in idle state of FSM), they get partially assigned
in this process, and condition where there is reset asserted and a rising edge
gets not defined properly.

Rework all such constructs as:

  if rising_edge(...) then
    ...
  end if;

  if reset then
    ...
  end if;

Which is correct, even if not all clocked registers have a reset clause.

This is particularely useful when we use records for grouping FSM state.